chore: move the file management components related into the file-management folder
